+++ KioChess/Engine/Models/Config/KingSafetyEvaluation.cs
@@ -6,7 +6,7 @@
         public int KingShieldPreFaceValue { get; set; }
         public int KingZoneOpenFileValue { get; set; }
         public double AttackValueFactor { get; set; }
-        public int[] PieceAttackValue { get; set; }
+        public short[] PieceAttackValue { get; set; }
         public double[] AttackWeight { get; set; }
     }
 }
